backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
moviepy==1.0.3